chore move user-components file
