+++ app/user/[id]/page-components/UserDetailedIntro.tsx
